requirements.txt
# Web framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
flask-jwt-extended==4.6.0

# Database
psycopg[binary,pool]>=3.2.0

# Configuration
python-dotenv==1.0.0

# Validation
pydantic>=2.9.0

# Graph algorithms
networkx==3.3

# Async support
eventlet==0.35.2

# Production server
gunicorn==21.2.0

# Testing
pytest==8.2.0
pytest-flask==1.3.0
pytest-cov==5.0.0